Python Mailer Util/requirements.txt
pyodbc==4.0.32
PyYAML==5.4.1